requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
datasets==2.19.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.42
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
Pygments==2.18.0
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
setuptools==68.2.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
ucimlrepo==0.0.7
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.41.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
